requirements/common.txt
Flask==1.1.2
requests==2.25.1
opentelemetry-instrumentation-flask==0.20b0
opentelemetry-instrumentation-requests==0.20b0

